Euro pose filter implementation
The “One Euro” pose filter does two things:

1. It’s an adaptive low-pass filter that decreases
lag / increases jitter when a joint is traveling fast
and does the opposite when a joint is traveling
slow, and

2. It makes it easy to tune the jitter/lag tradeoff
by changing the variables at the top of
VROPoseFilterEuro.cpp.

The values chosen for kEuroBeta and kEuroFCMin
were based on hand-tuning.

+++ ViroRenderer/VROBodyTracker.h
@@ -100,10 +100,10 @@ public:
     int getTileX() const { return _tileX; }
     int getTileY() const { return _tileY; }
     
-    void setCreationTime(float creationTimeMs) {
+    void setCreationTime(double creationTimeMs) {
         _creationTimeMs = creationTimeMs;
     }
-    float getCreationTime() const {
+    double getCreationTime() const {
         return _creationTimeMs;
     }
